+++ integration_test_apps/jason_only_app/lib/jason_only_app.ex
@@ -1,0 +1,2 @@
+defmodule JasonOnlyApp do
+end